cleanup spec
